client: move all json structs to `jsonmodels`
The `client` package had an indirect dependency on the `plugins`
package, which resulted in unwanted code being executed on `init`.
This made the `client` package unusable for third-party software.

This refactor moves all JSON structs to the `jsonmodels` package,
thus removing the unwanted dependency.
