+++ client/autopeering.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"net/http"
 
-	webapi_autopeering "github.com/iotaledger/goshimmer/plugins/webapi/autopeering"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 const (
@@ -13,8 +13,8 @@ const (
 
 // GetNeighbors gets the chosen/accepted neighbors.
 // If knownPeers is set, also all known peers to the node are returned additionally.
-func (api *GoShimmerAPI) GetNeighbors(knownPeers bool) (*webapi_autopeering.Response, error) {
-	res := &webapi_autopeering.Response{}
+func (api *GoShimmerAPI) GetNeighbors(knownPeers bool) (*jsonmodels.GetNeighborsResponse, error) {
+	res := &jsonmodels.GetNeighborsResponse{}
 	if err := api.do(http.MethodGet, func() string {
 		if !knownPeers {
 			return routeGetNeighbors
