+++ client/drng.go
@@ -3,7 +3,7 @@ package client
 import (
 	"net/http"
 
-	webapi_drng "github.com/iotaledger/goshimmer/plugins/webapi/drng"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 const (
@@ -14,9 +14,9 @@ const (
 
 // BroadcastCollectiveBeacon sends the given collective beacon (payload) by creating a message in the backend.
 func (api *GoShimmerAPI) BroadcastCollectiveBeacon(payload []byte) (string, error) {
-	res := &webapi_drng.CollectiveBeaconResponse{}
+	res := &jsonmodels.CollectiveBeaconResponse{}
 	if err := api.do(http.MethodPost, routeCollectiveBeacon,
-		&webapi_drng.CollectiveBeaconRequest{Payload: payload}, res); err != nil {
+		&jsonmodels.CollectiveBeaconRequest{Payload: payload}, res); err != nil {
 		return "", err
 	}
 
@@ -24,8 +24,8 @@ func (api *GoShimmerAPI) BroadcastCollectiveBeacon(payload []byte) (string, erro
 }
 
 // GetRandomness gets the current randomness.
-func (api *GoShimmerAPI) GetRandomness() (*webapi_drng.RandomnessResponse, error) {
-	res := &webapi_drng.RandomnessResponse{}
+func (api *GoShimmerAPI) GetRandomness() (*jsonmodels.RandomnessResponse, error) {
+	res := &jsonmodels.RandomnessResponse{}
 	if err := api.do(http.MethodGet, func() string {
 		return routeRandomness
 	}(), nil, res); err != nil {
@@ -35,8 +35,8 @@ func (api *GoShimmerAPI) GetRandomness() (*webapi_drng.RandomnessResponse, error
 }
 
 // GetCommittee gets the current committee.
-func (api *GoShimmerAPI) GetCommittee() (*webapi_drng.CommitteeResponse, error) {
-	res := &webapi_drng.CommitteeResponse{}
+func (api *GoShimmerAPI) GetCommittee() (*jsonmodels.CommitteeResponse, error) {
+	res := &jsonmodels.CommitteeResponse{}
 	if err := api.do(http.MethodGet, func() string {
 		return routeCommittee
 	}(), nil, res); err != nil {
