+++ client/spammer.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"net/http"
 
-	webapi_spammer "github.com/iotaledger/goshimmer/plugins/spammer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 const (
@@ -12,8 +12,8 @@ const (
 )
 
 // ToggleSpammer toggles the node internal spammer.
-func (api *GoShimmerAPI) ToggleSpammer(enable bool, mpm int) (*webapi_spammer.Response, error) {
-	res := &webapi_spammer.Response{}
+func (api *GoShimmerAPI) ToggleSpammer(enable bool, mpm int) (*jsonmodels.SpammerResponse, error) {
+	res := &jsonmodels.SpammerResponse{}
 	if err := api.do(http.MethodGet, func() string {
 		if enable {
 			return fmt.Sprintf("%s?cmd=start&mpm=%d", routeSpammer, mpm)
