+++ plugins/webapi/data/plugin.go
@@ -11,6 +11,7 @@ import (
 	"github.com/iotaledger/goshimmer/packages/tangle/payload"
 	"github.com/iotaledger/goshimmer/plugins/messagelayer"
 	"github.com/iotaledger/goshimmer/plugins/webapi"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // PluginName is the name of the web API data endpoint plugin.
@@ -39,26 +40,15 @@ func configure(plugin *node.Plugin) {
 // broadcastData creates a message of the given payload and
 // broadcasts it to the node's neighbors. It returns the message ID if successful.
 func broadcastData(c echo.Context) error {
-	var request Request
+	var request jsonmodels.DataRequest
 	if err := c.Bind(&request); err != nil {
 		log.Info(err.Error())
-		return c.JSON(http.StatusBadRequest, Response{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.DataResponse{Error: err.Error()})
 	}
 
 	msg, err := messagelayer.Tangle().IssuePayload(payload.NewGenericDataPayload(request.Data))
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, Response{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.DataResponse{Error: err.Error()})
 	}
-	return c.JSON(http.StatusOK, Response{ID: msg.ID().String()})
-}
-
-// Response contains the ID of the message sent.
-type Response struct {
-	ID    string `json:"id,omitempty"`
-	Error string `json:"error,omitempty"`
-}
-
-// Request contains the data of the message to send.
-type Request struct {
-	Data []byte `json:"data"`
+	return c.JSON(http.StatusOK, jsonmodels.DataResponse{ID: msg.ID().String()})
 }
