+++ plugins/webapi/drng/collectivebeacon.go
@@ -9,36 +9,26 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/drng"
 	"github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // collectiveBeaconHandler gets the current DRNG committee.
 func collectiveBeaconHandler(c echo.Context) error {
-	var request CollectiveBeaconRequest
+	var request jsonmodels.CollectiveBeaconRequest
 	if err := c.Bind(&request); err != nil {
 		log.Info(err.Error())
-		return c.JSON(http.StatusBadRequest, CollectiveBeaconResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.CollectiveBeaconResponse{Error: err.Error()})
 	}
 
 	marshalUtil := marshalutil.New(request.Payload)
 	parsedPayload, err := drng.CollectiveBeaconPayloadFromMarshalUtil(marshalUtil)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, CollectiveBeaconResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.CollectiveBeaconResponse{Error: err.Error()})
 	}
 
 	msg, err := messagelayer.Tangle().IssuePayload(parsedPayload)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, CollectiveBeaconResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.CollectiveBeaconResponse{Error: err.Error()})
 	}
-	return c.JSON(http.StatusOK, CollectiveBeaconResponse{ID: msg.ID().String()})
-}
-
-// CollectiveBeaconResponse is the HTTP response from broadcasting a collective beacon message.
-type CollectiveBeaconResponse struct {
-	ID    string `json:"id,omitempty"`
-	Error string `json:"error,omitempty"`
-}
-
-// CollectiveBeaconRequest is a request containing a collective beacon response.
-type CollectiveBeaconRequest struct {
-	Payload []byte `json:"payload"`
+	return c.JSON(http.StatusOK, jsonmodels.CollectiveBeaconResponse{ID: msg.ID().String()})
 }
