+++ plugins/webapi/drng/randomness.go
@@ -2,19 +2,19 @@ package drng
 
 import (
 	"net/http"
-	"time"
 
 	"github.com/labstack/echo"
 
 	"github.com/iotaledger/goshimmer/plugins/drng"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // randomnessHandler returns the current DRNG randomness used.
 func randomnessHandler(c echo.Context) error {
-	randomness := []Randomness{}
+	randomness := []jsonmodels.Randomness{}
 	for _, state := range drng.Instance().State {
 		randomness = append(randomness,
-			Randomness{
+			jsonmodels.Randomness{
 				InstanceID: state.Committee().InstanceID,
 				Round:      state.Randomness().Round,
 				Randomness: state.Randomness().Randomness,
@@ -22,21 +22,7 @@ func randomnessHandler(c echo.Context) error {
 			})
 	}
 
-	return c.JSON(http.StatusOK, RandomnessResponse{
+	return c.JSON(http.StatusOK, jsonmodels.RandomnessResponse{
 		Randomness: randomness,
 	})
-}
-
-// RandomnessResponse is the HTTP message containing the current DRNG randomness.
-type RandomnessResponse struct {
-	Randomness []Randomness `json:"randomness,omitempty"`
-	Error      string       `json:"error,omitempty"`
-}
-
-// Randomness defines the content of new randomness.
-type Randomness struct {
-	InstanceID uint32    `json:"instanceID,omitempty"`
-	Round      uint64    `json:"round,omitempty"`
-	Timestamp  time.Time `json:"timestamp,omitempty"`
-	Randomness []byte    `json:"randomness,omitempty"`
 }
