+++ plugins/webapi/jsonmodels/tangle.go
@@ -1,9 +1,7 @@
 package jsonmodels
 
 import (
-	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 	"github.com/iotaledger/goshimmer/packages/tangle"
-	"github.com/iotaledger/goshimmer/plugins/messagelayer"
 )
 
 // region Message ///////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -22,30 +20,6 @@ type Message struct {
 	TransactionID   string   `json:"transactionID,omitempty"`
 	Payload         []byte   `json:"payload"`
 	Signature       string   `json:"signature"`
-}
-
-// NewMessage returns a Message from the given tangle.Message.
-func NewMessage(message *tangle.Message) Message {
-	return Message{
-		ID:              message.ID().String(),
-		StrongParents:   message.StrongParents().ToStrings(),
-		WeakParents:     message.WeakParents().ToStrings(),
-		StrongApprovers: messagelayer.Tangle().Utils.ApprovingMessageIDs(message.ID(), tangle.StrongApprover).ToStrings(),
-		WeakApprovers:   messagelayer.Tangle().Utils.ApprovingMessageIDs(message.ID(), tangle.WeakApprover).ToStrings(),
-		IssuerPublicKey: message.IssuerPublicKey().String(),
-		IssuingTime:     message.IssuingTime().Unix(),
-		SequenceNumber:  message.SequenceNumber(),
-		PayloadType:     message.Payload().Type().String(),
-		TransactionID: func() string {
-			if message.Payload().Type() == ledgerstate.TransactionType {
-				return message.Payload().(*ledgerstate.Transaction).ID().Base58()
-			}
-
-			return ""
-		}(),
-		Payload:   message.Payload().Bytes(),
-		Signature: message.Signature().String(),
-	}
 }
 
 // endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
