+++ plugins/webapi/mana/allmana.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/mana"
 	manaPlugin "github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // getAllManaHandler handles the request.
@@ -16,7 +17,7 @@ func getAllManaHandler(c echo.Context) error {
 	t := time.Now()
 	access, tAccess, err := manaPlugin.GetManaMap(mana.AccessMana, t)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, GetAllManaResponse{
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetAllManaResponse{
 			Error: err.Error(),
 		})
 	}
@@ -26,7 +27,7 @@ func getAllManaHandler(c echo.Context) error {
 	})
 	consensus, tConsensus, err := manaPlugin.GetManaMap(mana.ConsensusMana, t)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, GetAllManaResponse{
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetAllManaResponse{
 			Error: err.Error(),
 		})
 	}
@@ -34,19 +35,10 @@ func getAllManaHandler(c echo.Context) error {
 	sort.Slice(consensusList, func(i, j int) bool {
 		return consensusList[i].Mana > consensusList[j].Mana
 	})
-	return c.JSON(http.StatusOK, GetAllManaResponse{
+	return c.JSON(http.StatusOK, jsonmodels.GetAllManaResponse{
 		Access:             accessList,
 		AccessTimestamp:    tAccess.Unix(),
 		Consensus:          consensusList,
 		ConsensusTimestamp: tConsensus.Unix(),
 	})
-}
-
-// GetAllManaResponse is the request to a getAllManaHandler request.
-type GetAllManaResponse struct {
-	Access             []mana.NodeStr `json:"access"`
-	AccessTimestamp    int64          `json:"accessTimestamp"`
-	Consensus          []mana.NodeStr `json:"consensus"`
-	ConsensusTimestamp int64          `json:"consensusTimestamp"`
-	Error              string         `json:"error,omitempty"`
 }
