+++ plugins/webapi/mana/logsHandler.go
@@ -10,19 +10,20 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/mana"
 	manaPlugin "github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // getEventLogsHandler handles the request.
 func getEventLogsHandler(c echo.Context) error {
-	var req GetEventLogsRequest
+	var req jsonmodels.GetEventLogsRequest
 	if err := c.Bind(&req); err != nil {
-		return c.JSON(http.StatusBadRequest, GetEventLogsResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetEventLogsResponse{Error: err.Error()})
 	}
 	var nodeIDs []identity.ID
 	for _, nodeID := range req.NodeIDs {
 		_nodeID, err := mana.IDFromStr(nodeID)
 		if err != nil {
-			return c.JSON(http.StatusBadRequest, GetEventLogsResponse{Error: err.Error()})
+			return c.JSON(http.StatusBadRequest, jsonmodels.GetEventLogsResponse{Error: err.Error()})
 		}
 		nodeIDs = append(nodeIDs, _nodeID)
 	}
@@ -33,14 +34,14 @@ func getEventLogsHandler(c echo.Context) error {
 		endTime = time.Now()
 	}
 	if endTime.Before(startTime) {
-		return c.JSON(http.StatusBadRequest, GetEventLogsResponse{Error: "time interval mismatch. endTime cannot be before startTime"})
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetEventLogsResponse{Error: "time interval mismatch. endTime cannot be before startTime"})
 	}
 	logs, err := manaPlugin.GetLoggedEvents(nodeIDs, startTime, endTime.Add(1*time.Second))
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, GetEventLogsResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetEventLogsResponse{Error: err.Error()})
 	}
 
-	res := make(map[string]*EventLogsJSON)
+	res := make(map[string]*jsonmodels.EventLogsJSON)
 	for ID, l := range logs {
 		var pledgesJSON []*mana.PledgedEventJSON
 		for _, p := range l.Pledge {
@@ -51,37 +52,16 @@ func getEventLogsHandler(c echo.Context) error {
 		for _, r := range l.Revoke {
 			revokesJSON = append(revokesJSON, r.ToJSONSerializable().(*mana.RevokedEventJSON))
 		}
-		eventsJSON := &EventLogsJSON{
+		eventsJSON := &jsonmodels.EventLogsJSON{
 			Pledge: pledgesJSON,
 			Revoke: revokesJSON,
 		}
 		res[base58.Encode(ID.Bytes())] = eventsJSON
 	}
 
-	return c.JSON(http.StatusOK, GetEventLogsResponse{
+	return c.JSON(http.StatusOK, jsonmodels.GetEventLogsResponse{
 		Logs:      res,
 		StartTime: startTime.Unix(),
 		EndTime:   endTime.Unix(),
 	})
-}
-
-// EventLogsJSON is a events log in JSON.
-type EventLogsJSON struct {
-	Pledge []*mana.PledgedEventJSON `json:"pledge"`
-	Revoke []*mana.RevokedEventJSON `json:"revoke"`
-}
-
-// GetEventLogsRequest is the request.
-type GetEventLogsRequest struct {
-	NodeIDs   []string `json:"nodeIDs"`
-	StartTime int64    `json:"startTime"`
-	EndTime   int64    `json:"endTime"`
-}
-
-// GetEventLogsResponse is the response.
-type GetEventLogsResponse struct {
-	Logs      map[string]*EventLogsJSON `json:"logs"`
-	Error     string                    `json:"error,omitempty"`
-	StartTime int64                     `json:"startTime"`
-	EndTime   int64                     `json:"endTime"`
 }
