+++ plugins/webapi/mana/mana.go
@@ -12,17 +12,18 @@ import (
 	"github.com/iotaledger/goshimmer/packages/mana"
 	"github.com/iotaledger/goshimmer/plugins/autopeering/local"
 	manaPlugin "github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // getManaHandler handles the request.
 func getManaHandler(c echo.Context) error {
-	var request GetManaRequest
+	var request jsonmodels.GetManaRequest
 	if err := c.Bind(&request); err != nil {
-		return c.JSON(http.StatusBadRequest, GetManaResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetManaResponse{Error: err.Error()})
 	}
 	ID, err := mana.IDFromStr(request.NodeID)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, GetManaResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetManaResponse{Error: err.Error()})
 	}
 	if request.NodeID == "" {
 		ID = local.GetInstance().ID()
@@ -34,7 +35,7 @@ func getManaHandler(c echo.Context) error {
 			accessMana = 0
 			tAccess = t
 		} else {
-			return c.JSON(http.StatusBadRequest, GetManaResponse{Error: err.Error()})
+			return c.JSON(http.StatusBadRequest, jsonmodels.GetManaResponse{Error: err.Error()})
 		}
 	}
 	consensusMana, tConsensus, err := manaPlugin.GetConsensusMana(ID, t)
@@ -43,11 +44,11 @@ func getManaHandler(c echo.Context) error {
 			consensusMana = 0
 			tConsensus = t
 		} else {
-			return c.JSON(http.StatusBadRequest, GetManaResponse{Error: err.Error()})
+			return c.JSON(http.StatusBadRequest, jsonmodels.GetManaResponse{Error: err.Error()})
 		}
 	}
 
-	return c.JSON(http.StatusOK, GetManaResponse{
+	return c.JSON(http.StatusOK, jsonmodels.GetManaResponse{
 		ShortNodeID:        ID.String(),
 		NodeID:             base58.Encode(ID.Bytes()),
 		Access:             accessMana,
@@ -55,20 +56,4 @@ func getManaHandler(c echo.Context) error {
 		Consensus:          consensusMana,
 		ConsensusTimestamp: tConsensus.Unix(),
 	})
-}
-
-// GetManaRequest is the request for get mana.
-type GetManaRequest struct {
-	NodeID string `json:"nodeID"`
-}
-
-// GetManaResponse defines the response for get mana.
-type GetManaResponse struct {
-	Error              string  `json:"error,omitempty"`
-	ShortNodeID        string  `json:"shortNodeID"`
-	NodeID             string  `json:"nodeID"`
-	Access             float64 `json:"access"`
-	AccessTimestamp    int64   `json:"accessTimestamp"`
-	Consensus          float64 `json:"consensus"`
-	ConsensusTimestamp int64   `json:"consensusTimestamp"`
 }
