+++ plugins/webapi/mana/nhighest.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/mana"
 	manaPlugin "github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // getNHighestAccessHandler handles a /mana/access/nhighest request.
@@ -24,25 +25,18 @@ func getNHighestConsensusHandler(c echo.Context) error {
 func nHighestHandler(c echo.Context, manaType mana.Type) error {
 	number, err := strconv.ParseUint(c.QueryParam("number"), 10, 32)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, GetNHighestResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetNHighestResponse{Error: err.Error()})
 	}
 	highestNodes, t, err := manaPlugin.GetHighestManaNodes(manaType, uint(number))
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, GetNHighestResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetNHighestResponse{Error: err.Error()})
 	}
 	var res []mana.NodeStr
 	for _, n := range highestNodes {
 		res = append(res, n.ToNodeStr())
 	}
-	return c.JSON(http.StatusOK, GetNHighestResponse{
+	return c.JSON(http.StatusOK, jsonmodels.GetNHighestResponse{
 		Nodes:     res,
 		Timestamp: t.Unix(),
 	})
-}
-
-// GetNHighestResponse holds info about nodes and their mana values.
-type GetNHighestResponse struct {
-	Error     string         `json:"error,omitempty"`
-	Nodes     []mana.NodeStr `json:"nodes,omitempty"`
-	Timestamp int64          `json:"timestamp"`
 }
