+++ plugins/webapi/mana/online.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/mana"
 	manaPlugin "github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 func getOnlineAccessHandler(c echo.Context) error {
@@ -22,11 +23,11 @@ func getOnlineConsensusHandler(c echo.Context) error {
 func getOnlineHandler(c echo.Context, manaType mana.Type) error {
 	onlinePeersMana, t, err := manaPlugin.GetOnlineNodes(manaType)
 	if err != nil {
-		return c.JSON(http.StatusNotFound, GetOnlineResponse{Error: err.Error()})
+		return c.JSON(http.StatusNotFound, jsonmodels.GetOnlineResponse{Error: err.Error()})
 	}
-	resp := make([]OnlineNodeStr, 0)
+	resp := make([]jsonmodels.OnlineNodeStr, 0)
 	for index, value := range onlinePeersMana {
-		resp = append(resp, OnlineNodeStr{
+		resp = append(resp, jsonmodels.OnlineNodeStr{
 			OnlineRank: index + 1,
 			ShortID:    value.ID.String(),
 			ID:         base58.Encode(value.ID.Bytes()),
@@ -34,23 +35,8 @@ func getOnlineHandler(c echo.Context, manaType mana.Type) error {
 		})
 	}
 
-	return c.JSON(http.StatusOK, GetOnlineResponse{
+	return c.JSON(http.StatusOK, jsonmodels.GetOnlineResponse{
 		Online:    resp,
 		Timestamp: t.Unix(),
 	})
-}
-
-// GetOnlineResponse is the response to an online mana request.
-type GetOnlineResponse struct {
-	Online    []OnlineNodeStr `json:"online"`
-	Error     string          `json:"error,omitempty"`
-	Timestamp int64           `json:"timestamp"`
-}
-
-// OnlineNodeStr holds information about online rank, nodeID and mana,
-type OnlineNodeStr struct {
-	OnlineRank int     `json:"rank"`
-	ShortID    string  `json:"shortNodeID"`
-	ID         string  `json:"nodeID"`
-	Mana       float64 `json:"mana"`
 }
