+++ plugins/webapi/mana/pastconsensusvectormetadata.go
@@ -5,25 +5,19 @@ import (
 
 	"github.com/labstack/echo"
 
-	"github.com/iotaledger/goshimmer/packages/mana"
 	manaPlugin "github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // getPastConsensusVectorMetadataHandler handles the request.
 func getPastConsensusVectorMetadataHandler(c echo.Context) error {
 	metadata := manaPlugin.GetPastConsensusManaVectorMetadata()
 	if metadata == nil {
-		return c.JSON(http.StatusOK, PastConsensusVectorMetadataResponse{
+		return c.JSON(http.StatusOK, jsonmodels.PastConsensusVectorMetadataResponse{
 			Error: "Past consensus mana vector metadata not found",
 		})
 	}
-	return c.JSON(http.StatusOK, PastConsensusVectorMetadataResponse{
+	return c.JSON(http.StatusOK, jsonmodels.PastConsensusVectorMetadataResponse{
 		Metadata: metadata,
 	})
-}
-
-// PastConsensusVectorMetadataResponse is the response.
-type PastConsensusVectorMetadataResponse struct {
-	Metadata *mana.ConsensusBasePastManaVectorMetadata `json:"metadata,omitempty"`
-	Error    string                                    `json:"error,omitempty"`
 }
