+++ plugins/webapi/mana/pending.go
@@ -7,35 +7,23 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 	manaPlugin "github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // GetPendingMana handles the request.
 func GetPendingMana(c echo.Context) error {
-	var req PendingRequest
+	var req jsonmodels.PendingRequest
 	if err := c.Bind(&req); err != nil {
-		return c.JSON(http.StatusBadRequest, PendingResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.PendingResponse{Error: err.Error()})
 	}
 	outputID, err := ledgerstate.OutputIDFromBase58(req.OutputID)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, PendingResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.PendingResponse{Error: err.Error()})
 	}
 	pending, t := manaPlugin.PendingManaOnOutput(outputID)
-	return c.JSON(http.StatusOK, PendingResponse{
+	return c.JSON(http.StatusOK, jsonmodels.PendingResponse{
 		Mana:      pending,
 		OutputID:  outputID.Base58(),
 		Timestamp: t.Unix(),
 	})
-}
-
-// PendingRequest is the pending mana request.
-type PendingRequest struct {
-	OutputID string `json:"outputID"`
-}
-
-// PendingResponse is the pending mana response.
-type PendingResponse struct {
-	Mana      float64 `json:"mana"`
-	OutputID  string  `json:"outputID"`
-	Error     string  `json:"error,omitempty"`
-	Timestamp int64   `json:"timestamp"`
 }
