+++ plugins/webapi/mana/percentile.go
@@ -11,17 +11,18 @@ import (
 	"github.com/iotaledger/goshimmer/packages/mana"
 	"github.com/iotaledger/goshimmer/plugins/autopeering/local"
 	manaPlugin "github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // getPercentileHandler handles the request.
 func getPercentileHandler(c echo.Context) error {
-	var request GetPercentileRequest
+	var request jsonmodels.GetPercentileRequest
 	if err := c.Bind(&request); err != nil {
-		return c.JSON(http.StatusBadRequest, GetPercentileResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetPercentileResponse{Error: err.Error()})
 	}
 	ID, err := mana.IDFromStr(request.NodeID)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, GetPercentileResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetPercentileResponse{Error: err.Error()})
 	}
 	if request.NodeID == "" {
 		ID = local.GetInstance().ID()
@@ -29,29 +30,29 @@ func getPercentileHandler(c echo.Context) error {
 	t := time.Now()
 	access, tAccess, err := manaPlugin.GetManaMap(mana.AccessMana, t)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, GetPercentileResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetPercentileResponse{Error: err.Error()})
 	}
 	accessPercentile, err := access.GetPercentile(ID)
 	if err != nil {
 		if xerrors.Is(err, mana.ErrNodeNotFoundInBaseManaVector) {
 			accessPercentile = 0
 		} else {
-			return c.JSON(http.StatusBadRequest, GetManaResponse{Error: err.Error()})
+			return c.JSON(http.StatusBadRequest, jsonmodels.GetManaResponse{Error: err.Error()})
 		}
 	}
 	consensus, tConsensus, err := manaPlugin.GetManaMap(mana.ConsensusMana, t)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, GetPercentileResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.GetPercentileResponse{Error: err.Error()})
 	}
 	consensusPercentile, err := consensus.GetPercentile(ID)
 	if err != nil {
 		if xerrors.Is(err, mana.ErrNodeNotFoundInBaseManaVector) {
 			consensusPercentile = 0
 		} else {
-			return c.JSON(http.StatusBadRequest, GetManaResponse{Error: err.Error()})
+			return c.JSON(http.StatusBadRequest, jsonmodels.GetManaResponse{Error: err.Error()})
 		}
 	}
-	return c.JSON(http.StatusOK, GetPercentileResponse{
+	return c.JSON(http.StatusOK, jsonmodels.GetPercentileResponse{
 		ShortNodeID:        ID.String(),
 		NodeID:             base58.Encode(ID.Bytes()),
 		Access:             accessPercentile,
@@ -59,20 +60,4 @@ func getPercentileHandler(c echo.Context) error {
 		Consensus:          consensusPercentile,
 		ConsensusTimestamp: tConsensus.Unix(),
 	})
-}
-
-// GetPercentileRequest is the request object of mana/percentile.
-type GetPercentileRequest struct {
-	NodeID string `json:"nodeID"`
-}
-
-// GetPercentileResponse holds info about the mana percentile(s) of a node.
-type GetPercentileResponse struct {
-	Error              string  `json:"error,omitempty"`
-	ShortNodeID        string  `json:"shortNodeID"`
-	NodeID             string  `json:"nodeID"`
-	Access             float64 `json:"access"`
-	AccessTimestamp    int64   `json:"accessTimestamp"`
-	Consensus          float64 `json:"consensus"`
-	ConsensusTimestamp int64   `json:"consensusTimestamp"`
 }
