+++ plugins/webapi/message/plugin.go
@@ -8,6 +8,7 @@ import (
 	"github.com/iotaledger/hive.go/node"
 	"github.com/labstack/echo"
 
+	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 	"github.com/iotaledger/goshimmer/packages/tangle"
 	"github.com/iotaledger/goshimmer/packages/tangle/payload"
 	"github.com/iotaledger/goshimmer/plugins/messagelayer"
@@ -50,7 +51,26 @@ func GetMessage(c echo.Context) (err error) {
 	}
 
 	if messagelayer.Tangle().Storage.Message(messageID).Consume(func(message *tangle.Message) {
-		err = c.JSON(http.StatusOK, jsonmodels.NewMessage(message))
+		err = c.JSON(http.StatusOK, jsonmodels.Message{
+			ID:              message.ID().String(),
+			StrongParents:   message.StrongParents().ToStrings(),
+			WeakParents:     message.WeakParents().ToStrings(),
+			StrongApprovers: messagelayer.Tangle().Utils.ApprovingMessageIDs(message.ID(), tangle.StrongApprover).ToStrings(),
+			WeakApprovers:   messagelayer.Tangle().Utils.ApprovingMessageIDs(message.ID(), tangle.WeakApprover).ToStrings(),
+			IssuerPublicKey: message.IssuerPublicKey().String(),
+			IssuingTime:     message.IssuingTime().Unix(),
+			SequenceNumber:  message.SequenceNumber(),
+			PayloadType:     message.Payload().Type().String(),
+			TransactionID: func() string {
+				if message.Payload().Type() == ledgerstate.TransactionType {
+					return message.Payload().(*ledgerstate.Transaction).ID().Base58()
+				}
+
+				return ""
+			}(),
+			Payload:   message.Payload().Bytes(),
+			Signature: message.Signature().String(),
+		})
 	}) {
 		return
 	}
