+++ plugins/webapi/tools/message/pastcone.go
@@ -9,19 +9,20 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/tangle"
 	"github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // PastconeHandler process a pastcone request.
 func PastconeHandler(c echo.Context) error {
 	var checkedMessageCount int
-	var request PastconeRequest
+	var request jsonmodels.PastconeRequest
 	if err := c.Bind(&request); err != nil {
-		return c.JSON(http.StatusBadRequest, PastconeResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.PastconeResponse{Error: err.Error()})
 	}
 
 	msgID, err := tangle.NewMessageID(request.ID)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, PastconeResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.PastconeResponse{Error: err.Error()})
 	}
 
 	// create a new stack that hold messages to check
@@ -44,7 +45,7 @@ func PastconeHandler(c echo.Context) error {
 		msgMetadataObject := messagelayer.Tangle().Storage.MessageMetadata(currentMsgID)
 
 		if !msgObject.Exists() || !msgMetadataObject.Exists() {
-			return c.JSON(http.StatusOK, PastconeResponse{Exist: false, PastConeSize: checkedMessageCount, Error: fmt.Sprintf("couldn't find %s message on node", currentMsgID)})
+			return c.JSON(http.StatusOK, jsonmodels.PastconeResponse{Exist: false, PastConeSize: checkedMessageCount, Error: fmt.Sprintf("couldn't find %s message on node", currentMsgID)})
 		}
 
 		// get parent1 and parent2
@@ -74,18 +75,5 @@ func PastconeHandler(c echo.Context) error {
 		msgObject.Release()
 		msgMetadataObject.Release()
 	}
-	return c.JSON(http.StatusOK, PastconeResponse{Exist: true, PastConeSize: checkedMessageCount})
-}
-
-// PastconeRequest holds the message id to query.
-type PastconeRequest struct {
-	ID string `json:"id"`
-}
-
-// PastconeResponse is the HTTP response containing the number of messages in the past cone and if all messages of the past cone
-// exist on the node.
-type PastconeResponse struct {
-	Exist        bool   `json:"exist,omitempty"`
-	PastConeSize int    `json:"pastConeSize,omitempty"`
-	Error        string `json:"error,omitempty"`
+	return c.JSON(http.StatusOK, jsonmodels.PastconeResponse{Exist: true, PastConeSize: checkedMessageCount})
 }
