+++ plugins/webapi/value/attachments.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 	"github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels/value"
 
 	"github.com/labstack/echo"
 	"github.com/labstack/gommon/log"
@@ -15,18 +16,18 @@ func attachmentsHandler(c echo.Context) error {
 	txnID, err := ledgerstate.TransactionIDFromBase58(c.QueryParam("txnID"))
 	if err != nil {
 		log.Info(err)
-		return c.JSON(http.StatusBadRequest, AttachmentsResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, value.AttachmentsResponse{Error: err.Error()})
 	}
 
-	var valueObjs []ValueObject
+	var valueObjs []value.ValueObject
 
 	// get txn by txn id
 	tx := messagelayer.Tangle().LedgerState.Transaction(txnID)
 	defer tx.Release()
 	if !tx.Exists() {
-		return c.JSON(http.StatusNotFound, AttachmentsResponse{Error: "Transaction not found"})
+		return c.JSON(http.StatusNotFound, value.AttachmentsResponse{Error: "Transaction not found"})
 	}
-	txn := ParseTransaction(tx.Unwrap())
+	txn := value.ParseTransaction(tx.Unwrap())
 
 	// get attachments by txn id
 	for _, attachmentObj := range messagelayer.Tangle().Storage.Attachments(txnID) {
@@ -48,25 +49,12 @@ func attachmentsHandler(c echo.Context) error {
 			parents = append(parents, parent.String())
 		}
 		// append value object
-		valueObjs = append(valueObjs, ValueObject{
+		valueObjs = append(valueObjs, value.ValueObject{
 			ID:          payload.ID().String(),
 			Parents:     parents,
 			Transaction: txn,
 		})
 	}
 
-	return c.JSON(http.StatusOK, AttachmentsResponse{Attachments: valueObjs})
-}
-
-// AttachmentsResponse is the HTTP response from retrieving value objects.
-type AttachmentsResponse struct {
-	Attachments []ValueObject `json:"attachments,omitempty"`
-	Error       string        `json:"error,omitempty"`
-}
-
-// ValueObject holds the information of a value object.
-type ValueObject struct {
-	ID          string      `json:"id"`
-	Parents     []string    `json:"parents"`
-	Transaction Transaction `json:"transaction"`
+	return c.JSON(http.StatusOK, value.AttachmentsResponse{Attachments: valueObjs})
 }
