+++ plugins/webapi/value/gettransactionbyid.go
@@ -7,42 +7,43 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 	"github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels/value"
 )
 
 // getTransactionByIDHandler gets the transaction by id.
 func getTransactionByIDHandler(c echo.Context) error {
 	txID, err := ledgerstate.TransactionIDFromBase58(c.QueryParam("txnID"))
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, GetTransactionByIDResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, value.GetTransactionByIDResponse{Error: err.Error()})
 	}
 
 	// get txn by txn id
 	cachedTxnMetaObj := messagelayer.Tangle().LedgerState.TransactionMetadata(txID)
 	defer cachedTxnMetaObj.Release()
 	if !cachedTxnMetaObj.Exists() {
-		return c.JSON(http.StatusNotFound, GetTransactionByIDResponse{Error: "Transaction not found"})
+		return c.JSON(http.StatusNotFound, value.GetTransactionByIDResponse{Error: "Transaction not found"})
 	}
 	cachedTxnObj := messagelayer.Tangle().LedgerState.Transaction(txID)
 	defer cachedTxnObj.Release()
 	if !cachedTxnObj.Exists() {
-		return c.JSON(http.StatusNotFound, GetTransactionByIDResponse{Error: "Transaction not found"})
+		return c.JSON(http.StatusNotFound, value.GetTransactionByIDResponse{Error: "Transaction not found"})
 	}
-	txn := ParseTransaction(cachedTxnObj.Unwrap())
+	txn := value.ParseTransaction(cachedTxnObj.Unwrap())
 
 	txMetadata := cachedTxnMetaObj.Unwrap()
 	txInclusionState, err := messagelayer.Tangle().LedgerState.TransactionInclusionState(txID)
 	if err != nil {
-		return c.JSON(http.StatusOK, GetTransactionByIDResponse{Error: err.Error()})
+		return c.JSON(http.StatusOK, value.GetTransactionByIDResponse{Error: err.Error()})
 	}
 	cachedBranch := messagelayer.Tangle().LedgerState.BranchDAG.Branch(txMetadata.BranchID())
 	defer cachedBranch.Release()
 	if !cachedTxnObj.Exists() {
-		return c.JSON(http.StatusNotFound, GetTransactionByIDResponse{Error: "Branch not found"})
+		return c.JSON(http.StatusNotFound, value.GetTransactionByIDResponse{Error: "Branch not found"})
 	}
 	branch := cachedBranch.Unwrap()
 
-	return c.JSON(http.StatusOK, GetTransactionByIDResponse{
-		TransactionMetadata: TransactionMetadata{
+	return c.JSON(http.StatusOK, value.GetTransactionByIDResponse{
+		TransactionMetadata: value.TransactionMetadata{
 			BranchID:           branch.ID().String(),
 			Solid:              txMetadata.Solid(),
 			SolidificationTime: txMetadata.SolidificationTime().Unix(),
@@ -50,7 +51,7 @@ func getTransactionByIDHandler(c echo.Context) error {
 			LazyBooked:         txMetadata.LazyBooked(),
 		},
 		Transaction: txn,
-		InclusionState: InclusionState{
+		InclusionState: value.InclusionState{
 			Confirmed:   txInclusionState == ledgerstate.Confirmed,
 			Conflicting: messagelayer.Tangle().LedgerState.TransactionConflicting(txID),
 			Liked:       branch.Liked(),
@@ -60,12 +61,4 @@ func getTransactionByIDHandler(c echo.Context) error {
 			Preferred:   false,
 		},
 	})
-}
-
-// GetTransactionByIDResponse is the HTTP response from retrieving transaction.
-type GetTransactionByIDResponse struct {
-	TransactionMetadata TransactionMetadata `json:"transactionMetadata"`
-	Transaction         Transaction         `json:"transaction"`
-	InclusionState      InclusionState      `json:"inclusion_state"`
-	Error               string              `json:"error,omitempty"`
 }
