+++ plugins/webapi/value/sendtransaction.go
@@ -14,7 +14,10 @@ import (
 	"github.com/iotaledger/goshimmer/packages/mana"
 	"github.com/iotaledger/goshimmer/packages/tangle"
 	"github.com/iotaledger/goshimmer/plugins/messagelayer"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels/value"
 )
+
+const maxBookedAwaitTime = 5 * time.Second
 
 var (
 	sendTxMu sync.Mutex
@@ -27,22 +30,22 @@ func sendTransactionHandler(c echo.Context) error {
 	sendTxMu.Lock()
 	defer sendTxMu.Unlock()
 
-	var request SendTransactionRequest
+	var request value.SendTransactionRequest
 	if err := c.Bind(&request); err != nil {
-		return c.JSON(http.StatusBadRequest, SendTransactionResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, value.SendTransactionResponse{Error: err.Error()})
 	}
 
 	// parse tx
 	tx, _, err := ledgerstate.TransactionFromBytes(request.TransactionBytes)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, SendTransactionResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, value.SendTransactionResponse{Error: err.Error()})
 	}
 
 	// validate allowed mana pledge nodes.
 	allowedAccessMana := messagelayer.GetAllowedPledgeNodes(mana.AccessMana)
 	if allowedAccessMana.IsFilterEnabled {
 		if !allowedAccessMana.Allowed.Has(tx.Essence().AccessPledgeID()) {
-			return c.JSON(http.StatusBadRequest, SendTransactionResponse{
+			return c.JSON(http.StatusBadRequest, value.SendTransactionResponse{
 				Error: fmt.Errorf("not allowed to pledge access mana to %s: %w", tx.Essence().AccessPledgeID().String(), ErrNotAllowedToPledgeManaToNode).Error(),
 			})
 		}
@@ -50,7 +53,7 @@ func sendTransactionHandler(c echo.Context) error {
 	allowedConsensusMana := messagelayer.GetAllowedPledgeNodes(mana.ConsensusMana)
 	if allowedConsensusMana.IsFilterEnabled {
 		if !allowedConsensusMana.Allowed.Has(tx.Essence().ConsensusPledgeID()) {
-			return c.JSON(http.StatusBadRequest, SendTransactionResponse{
+			return c.JSON(http.StatusBadRequest, value.SendTransactionResponse{
 				Error: fmt.Errorf("not allowed to pledge consensus mana to %s: %w", tx.Essence().ConsensusPledgeID().String(), ErrNotAllowedToPledgeManaToNode).Error(),
 			})
 		}
@@ -58,12 +61,12 @@ func sendTransactionHandler(c echo.Context) error {
 
 	// check transaction validity
 	if valid, err := messagelayer.Tangle().LedgerState.CheckTransaction(tx); !valid {
-		return c.JSON(http.StatusBadRequest, SendTransactionResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, value.SendTransactionResponse{Error: err.Error()})
 	}
 
 	// check if transaction is too old
 	if tx.Essence().Timestamp().Before(clock.SyncedTime().Add(-tangle.MaxReattachmentTimeMin)) {
-		return c.JSON(http.StatusBadRequest, SendTransactionResponse{Error: fmt.Sprintf("transaction timestamp is older than MaxReattachmentTime (%s) and cannot be issued", tangle.MaxReattachmentTimeMin)})
+		return c.JSON(http.StatusBadRequest, value.SendTransactionResponse{Error: fmt.Sprintf("transaction timestamp is older than MaxReattachmentTime (%s) and cannot be issued", tangle.MaxReattachmentTimeMin)})
 	}
 
 	// if transaction is in the future we wait until the time arrives
@@ -76,18 +79,7 @@ func sendTransactionHandler(c echo.Context) error {
 	}
 
 	if _, err := messagelayer.AwaitMessageToBeBooked(issueTransaction, tx.ID(), maxBookedAwaitTime); err != nil {
-		return c.JSON(http.StatusBadRequest, SendTransactionResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, value.SendTransactionResponse{Error: err.Error()})
 	}
-	return c.JSON(http.StatusOK, SendTransactionResponse{TransactionID: tx.ID().Base58()})
-}
-
-// SendTransactionRequest holds the transaction object(bytes) to send.
-type SendTransactionRequest struct {
-	TransactionBytes []byte `json:"txn_bytes"`
-}
-
-// SendTransactionResponse is the HTTP response from sending transaction.
-type SendTransactionResponse struct {
-	TransactionID string `json:"transaction_id,omitempty"`
-	Error         string `json:"error,omitempty"`
+	return c.JSON(http.StatusOK, value.SendTransactionResponse{TransactionID: tx.ID().Base58()})
 }
