+++ tools/integration-tests/tester/framework/peer.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/iotaledger/goshimmer/client"
 	walletseed "github.com/iotaledger/goshimmer/client/wallet/packages/seed"
-	"github.com/iotaledger/goshimmer/plugins/webapi/autopeering"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // Peer represents a GoShimmer node inside the Docker network
@@ -29,8 +29,8 @@ type Peer struct {
 	// Seed
 	*walletseed.Seed
 
-	chosen   []autopeering.Neighbor
-	accepted []autopeering.Neighbor
+	chosen   []jsonmodels.Neighbor
+	accepted []jsonmodels.Neighbor
 }
 
 // newPeer creates a new instance of Peer with the given information.
@@ -62,10 +62,10 @@ func (p *Peer) TotalNeighbors() int {
 }
 
 // SetNeighbors sets the neighbors of the peer accordingly.
-func (p *Peer) SetNeighbors(chosen, accepted []autopeering.Neighbor) {
-	p.chosen = make([]autopeering.Neighbor, len(chosen))
+func (p *Peer) SetNeighbors(chosen, accepted []jsonmodels.Neighbor) {
+	p.chosen = make([]jsonmodels.Neighbor, len(chosen))
 	copy(p.chosen, chosen)
 
-	p.accepted = make([]autopeering.Neighbor, len(accepted))
+	p.accepted = make([]jsonmodels.Neighbor, len(accepted))
 	copy(p.accepted, accepted)
 }
