+++ tools/integration-tests/tester/tests/consensus/consensus_conflicts_test.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/iotaledger/goshimmer/client/wallet/packages/seed"
 	"github.com/iotaledger/goshimmer/packages/ledgerstate"
-	"github.com/iotaledger/goshimmer/plugins/webapi/value"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels/value"
 	"github.com/iotaledger/goshimmer/tools/integration-tests/tester/framework"
 	"github.com/iotaledger/goshimmer/tools/integration-tests/tester/tests"
 )
@@ -23,7 +23,6 @@ import (
 // then issues valid value objects spending the genesis in both, deletes the partitions (and lets them merge)
 // and then checks that the conflicts are resolved via FPC.
 func TestConsensusFiftyFiftyOpinionSplit(t *testing.T) {
-
 	// override avg. network delay to accustom integration test slowness
 	backupFCoBAvgNetworkDelay := framework.ParaFCoBAverageNetworkDelay
 	backupBootstrapOnEveryNode := framework.ParaSyncBeaconOnEveryNode
@@ -78,7 +77,7 @@ func TestConsensusFiftyFiftyOpinionSplit(t *testing.T) {
 	// issue one transaction per peer to pledge mana to nodes
 	// leave one unspent output from splitting genesis transaction for further conflict creation
 
-	//prepare all the pledgingTxs
+	// prepare all the pledgingTxs
 	pledgingTxs := make([]*ledgerstate.Transaction, numberOfPeers+1)
 	pledgeSeed := make([]*seed.Seed, numberOfPeers+1)
 	receiverId := 0
@@ -107,7 +106,7 @@ func TestConsensusFiftyFiftyOpinionSplit(t *testing.T) {
 	log.Printf("waiting 2 * %d seconds avg. network delay + 5s to make the transactions confirmed", framework.ParaFCoBAverageNetworkDelay)
 	time.Sleep(time.Duration(framework.ParaFCoBAverageNetworkDelay)*2*time.Second + 5*time.Second)
 
-	//prepare two conflicting transactions from one additional unused genesis output
+	// prepare two conflicting transactions from one additional unused genesis output
 	conflictingTxs := make([]*ledgerstate.Transaction, len(n.Partitions()))
 	conflictingTxIDs := make([]string, len(n.Partitions()))
 	receiverSeeds := make([]*seed.Seed, len(n.Partitions()))
@@ -201,7 +200,6 @@ func TestConsensusFiftyFiftyOpinionSplit(t *testing.T) {
 	confirmed := make([]int, 2)
 
 	for i, conflictingTx := range conflictingTxIDs {
-
 		for _, p := range n.Peers() {
 			tx, err := p.GetTransactionByID(conflictingTx)
 			assert.NoError(t, err)
@@ -219,7 +217,6 @@ func TestConsensusFiftyFiftyOpinionSplit(t *testing.T) {
 	assert.Equal(t, len(n.Peers()), rejected[1], "the rejected count for second transaction should be equal to %d", len(n.Peers()))
 	assert.Equal(t, 0, confirmed[1], "the confirmed count for second transaction should be equal to 0")
 	assert.Equal(t, len(n.Peers()), confirmed[0], "the confirmed count for first transaction should be equal to the amount of peers %d", len(n.Peers()))
-
 }
 
 func CreateOutputs(input *ledgerstate.UTXOInput, inputBalance uint64, kp *ed25519.KeyPair, nOutputs int, pledgeID identity.ID, balanceType string) (*ledgerstate.Transaction, *seed.Seed) {
@@ -282,7 +279,7 @@ func createBalances(balanceType string, nOutputs int, inputBalance uint64) []uin
 			}
 			lastBalance, _ := ledgerstate.SafeSubUint64(remainingBalance, totalBalance)
 			outputBalances = append(outputBalances, lastBalance)
-			//outputBalances = append(outputBalances, createBalances("equal", nOutputs-1, remainingBalance)...)
+			// outputBalances = append(outputBalances, createBalances("equal", nOutputs-1, remainingBalance)...)
 			fmt.Printf("rady %v", outputBalances)
 		}
 	}
